fix(server): GMV 求和不再溢出 Postgres int4
/api/admin/stats、/api/merchant/stats、/api/orders/stats 里
sum(total_fen) 都强转成 ::int，Postgres int4 上限约 21.4 万元。
线上今天订单总额已经超过这个上限（业务本身允许单笔订单达到
50 件 * 99 份 * 单价上限），导致管理端统计页 500。改成 ::float8。

+++ server/src/routes/adminStats.ts
@@ -36,10 +36,10 @@ export const adminStatsRoutes = new Hono().get('/stats', requireAdmin, async (c)
     db
       .select({
         totalOrders: sql<number>`count(*)::int`,
-        gmvFen: sql<number>`COALESCE(sum(${orders.totalFen}), 0)::int`,
+        gmvFen: sql<number>`COALESCE(sum(${orders.totalFen}), 0)::float8`,
         completedOrders: sql<number>`count(*) FILTER (WHERE ${orders.status} = 'completed')::int`,
         todayOrders: sql<number>`count(*) FILTER (WHERE ${orders.createdAt} >= date_trunc('day', now()))::int`,
-        todayGmvFen: sql<number>`COALESCE(sum(${orders.totalFen}) FILTER (WHERE ${orders.createdAt} >= date_trunc('day', now())), 0)::int`,
+        todayGmvFen: sql<number>`COALESCE(sum(${orders.totalFen}) FILTER (WHERE ${orders.createdAt} >= date_trunc('day', now())), 0)::float8`,
       })
       .from(orders)
       .then((rows) => rows[0]),
@@ -75,7 +75,7 @@ export const adminStatsRoutes = new Hono().get('/stats', requireAdmin, async (c)
         bgColor: restaurants.bgColor,
         category: restaurants.category,
         orderCount: sql<number>`count(${orders.id})::int`,
-        gmvFen: sql<number>`COALESCE(sum(${orders.totalFen}), 0)::int`,
+        gmvFen: sql<number>`COALESCE(sum(${orders.totalFen}), 0)::float8`,
       })
       .from(restaurants)
       .leftJoin(orders, sql`${orders.restaurantId} = ${restaurants.id}`)
@@ -91,7 +91,7 @@ export const adminStatsRoutes = new Hono().get('/stats', requireAdmin, async (c)
       order_agg AS (
         SELECT date_trunc('day', created_at)::date AS day,
                count(*)::int AS order_count,
-               COALESCE(sum(total_fen), 0)::int AS gmv_fen
+               COALESCE(sum(total_fen), 0)::float8 AS gmv_fen
         FROM orders
         WHERE created_at >= date_trunc('day', now()) - interval '29 days'
         GROUP BY 1
